Rework output printing flags
The stdout and stderr flags are now orthogonal and no longer mutually
exclusive.

The `verbose` flag has been removed.  Specifying it either resulted in
an error or had no effect.

The `--status` flag has been removed.  It is no longer possible to print
out the status of the jobs.  If the `--output` option is given the
status will be saved to a file and can be retrieved from there.