Simple support for command namespaces
Setting the environment variable `FLIGHT_ACTION_NAMESPACE` will only
show actions with a name starting with
`"#{ENV['FLIGHT_ACTION_NAMESPACE']}-"`.  That prefix will also be
removed when defining the name of the command.

The name of the program in its help output is changed from
`flight-action` to `"flight-#{ENV['FLIGHT_ACTION_NAMESPACE']}"`.

Also the description of the program can be set from the environment
variable `FLIGHT_ACTION_DESCRIPTION`.